Refactor implicit priors to their own module
