+++ cuqi/distribution/__init__.py
@@ -5,7 +5,6 @@ from ._cmrf import CMRF
 from ._gamma import Gamma
 from ._gaussian import Gaussian, JointGaussianSqrtPrec
 from ._gmrf import GMRF
-from ._implicit_regularized_gaussian import ImplicitRegularizedGaussian, ImplicitRegularizedGMRF
 from ._inverse_gamma import InverseGamma
 from ._lmrf import LMRF
 from ._laplace import Laplace
