+++ cuqi/implicitprior/__init__.py
@@ -1,0 +1,2 @@
+from ._regularizedGaussian import RegularizedGaussian
+from ._regularizedGMRF import RegularizedGMRF
